back-end/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
psycopg==3.1.9
psycopg-binary==3.1.9
Pygments==2.15.1
python-dotenv==1.0.0
pytz==2023.3
redgreenunittest==0.1.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.1.0
